ft(first-feedbacks-ch1ch2): chore of Feedbacks on Challenge 1&2
with this chore the following adjustments are going to be implemented:

on the UI:
have the same CSS style when adding and when editing my diary entries
the date on the view-entry page should be placed at the bottom, not in the middle of the title and the description of the entry.

On APIs:
user should also have user credentials: first name, last name and email in the response spec. for sign up and sign in and add the userId on the entry APIs
the authorization token of the user should contain the user_id not the user_email.
When adding an entry, the date should be the date of creation, and this should be different to the editedOn date or time.
Implement the adjustment of the status code and the error message on THE DELETE API using a malformed token is fitting and should be: you are forbidden to perform this action! and the status code should be 401
Also implement the modification of the view-all API endpoint to include pagination

 [Delivers #169417227]